test(scan): separate qrcode worker tests
These tests weren't testing `Interpreter`, so they don't need to be in this file.
